xfs/111: fix golden output
In commit f4eee5126, this test was changed to run xfs_fsr to
bulkstat its way through the filesystem until it finds the corrupt
inode.  The golden output was changed to capture xfs_fsr's output,
but neglects the fact that when fsr's output is not a tty, all the
status and error messages are sent to syslog, not stdout.
Therefore, this test consistently fails because it expects output of
"$SCRATCH_MNT start inode=0" but this never appears.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Reviewed-by: Eryu Guan <guaneryu@gmail.com>
Signed-off-by: Eryu Guan <guaneryu@gmail.com>